Added renderer parameters.
The renderer classes can now take in parameters to init some values
like the window sizes and scale factor.

+++ engine/src/SDL/SDLTexture.cpp
@@ -22,7 +22,7 @@ SDLTexture::SDLTexture(ARenderer* renderer, SDL_Surface* fromSurface) : SDLTextu
     this->surf = fromSurface;
     this->tex = this->SdlRenderer->BuildTexture(fromSurface);
 
-    this->textureSize = FSize(fromSurface->w, fromSurface->h);
+    this->textureSize = FSize((float)fromSurface->w, (float)fromSurface->h);
     isDirty = false;
 }
